+++ CHANGELOG.md
@@ -6,10 +6,16 @@
 
 - Changelog added. (wiheto)
 - Preliminary BIDS operations that generate edge df from BIDS layout `edges_from_bids()` (wiheto)
+- Connectivity matrix as view option (wiheto)
+- Rotation of connectivity matrix (wiheto)
+- cm_order
+- Dual colours for positive and negative edges (wiheto)
 
 ### Update
 - Updated README with network neuroscience reference (wiheto)
 
 ### Fixed
 - Renaming "legend_tick_fontsize" and "legend_title_fontsize" to be more consistent (wiheto)
+- Renaming "highlightlevel" to "highlight_level" to be more consistent (wiheto)
 - Correcting some kwarg names in documentation (wiheto)
+- A bug with spring layout and node_color when only subset of nodes plotted